0-linear

+++ 0x1E-search_algorithms/0-linear.c
@@ -11,10 +11,10 @@
 
 int linear_search(int *array, size_t size, int value)
 {
-	size_t i;
-	for (i = 0; i <= size; i++)
+	int i;
+	for (i = 0; i <= (int)size; i++)
 	{
-		printf("Value checked array[%ld] = [%d]\n", i, array[i]);
+		printf("Value checked array[%d] = [%d]\n", i, array[i]);
 		if (array[i] == value)
 		{
 			return (i);
